% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_2ggplot2.R
\name{add_2ggplots}
\alias{add_2ggplots}
\title{Add two ggplots into a document object}
\usage{
add_2ggplots(mydoc, title = "Two plots", plot1, plot2, width = 3,
  height = 2.5)
}
\arguments{
\item{mydoc}{A document object}

\item{title}{An character string as a plot title}

\item{plot1}{An R code encoding the first ggplot}

\item{plot2}{An R code encoding the second ggplot}

\item{width}{plot width in inches}

\item{height}{plot height in inches}
}
\value{
a document object
}
\description{
Add two ggplots into a document object
}
\examples{
require(ggplot2)
require(magrittr)
require(officer)
gg1 <- ggplot(data = iris, aes(Sepal.Length, Petal.Length)) + geom_point()
gg2 <- ggplot(data = iris, aes(Sepal.Length, Petal.Length, color = Species)) + geom_point()
read_docx() \%>\% add_2ggplots(title="Two plots",plot1=gg1,plot2=gg2)
}
