% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_2ggplot2.R
\name{add_2plots}
\alias{add_2plots}
\title{Add two plots into a document object}
\usage{
add_2plots(mydoc, plotstring1, plotstring2, title = "", width = 3,
  height = 2.5, echo = FALSE)
}
\arguments{
\item{mydoc}{A document object}

\item{plotstring1}{An R code string encoding the first plot}

\item{plotstring2}{An R code string encoding the second plot}

\item{title}{An character string as a plot title}

\item{width}{plot width in inches}

\item{height}{plot height in inches}

\item{echo}{logical Whether or not show R code}
}
\value{
a document object
}
\description{
Add two plots into a document object
}
\examples{
require(magrittr)
require(officer)
plotstring1="plot(1:10)"
plotstring2="hist(rnorm(100))"
read_pptx() \%>\% add_2plots(plotstring1,plotstring2,title="Two plots") \%>\%
add_2plots(plotstring1,plotstring2,title="Two plots",echo=TRUE) \%>\% print(target="demo.pptx")
}
