% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flextable.R
\name{add_ggplot}
\alias{add_ggplot}
\title{Add ggplot into a document object}
\usage{
add_ggplot(mydoc, gg, title = "")
}
\arguments{
\item{mydoc}{A document object}

\item{gg}{An R code encoding a ggplot}

\item{title}{An character string as a plot title}
}
\value{
a document object
}
\description{
Add ggplot into a document object
}
\examples{
require(webr)
require(ggplot2)
require(officer)
require(magrittr)
gg <- ggplot(mtcars, aes(x = mpg , y = wt, colour = factor(am))) + geom_point()
read_pptx() \%>\% add_ggplot(gg)
read_docx() \%>\% add_ggplot(gg)
}
