% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flextable.R
\name{add_img}
\alias{add_img}
\title{Add plot into a document object}
\usage{
add_img(mydoc, plotstring, title = "", width = 7, height = 5,
  units = "in", res = 300, format = "emf", echo = FALSE, ...)
}
\arguments{
\item{mydoc}{A document object}

\item{plotstring}{An string of R code encoding plot}

\item{title}{An character string as a plot title}

\item{width}{the width of the device.}

\item{height}{the height of the device.}

\item{units}{The units in which height and width are given. Can be px (pixels, the default), in (inches), cm or mm.}

\item{res}{The nominal resolution in ppi which will be recorded in the bitmap file, if a positive integer. Also used for units other than the default, and to convert points to pixels.}

\item{format}{plot format}

\item{echo}{logical Whether or not show R code}

\item{...}{additional arguments passed to png()}
}
\value{
a document object
}
\description{
Add plot into a document object
}
\examples{
require(officer)
require(webr)
require(magrittr)
require(flextable)
read_pptx() \%>\% add_img("plot(mtcars)",format="png",res=300)
}
