% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flextable.R
\name{add_plot}
\alias{add_plot}
\title{Add plot into a document object}
\usage{
add_plot(mydoc, plotstring, title = "", vector = TRUE)
}
\arguments{
\item{mydoc}{A document object}

\item{plotstring}{String of an R code encoding a plot}

\item{title}{An character string as a plot title}

\item{vector}{A logical. If TRUE, vector graphics are produced instead, PNG images if FALSE.}
}
\value{
a document object
}
\description{
Add plot into a document object
}
\examples{
require(webr)
require(officer)
require(rvg)
require(magrittr)
read_pptx() \%>\% add_plot("plot(iris)")
read_docx() \%>\% add_plot("plot(iris)")
}
