% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flextable.R
\name{add_Rcode}
\alias{add_Rcode}
\title{Make a R code slide into a document object}
\usage{
add_Rcode(mydoc, code, title = "", preprocessing = "")
}
\arguments{
\item{mydoc}{A document object}

\item{code}{A character string encoding R codes}

\item{title}{An character string as a plot title}

\item{preprocessing}{A character string of R code as a preprocessing}
}
\value{
a document object
}
\description{
Make a R code slide into a document object
}
\examples{
#library(webr)
#library(magrittr)
#library(officer)
#code="fit=lm(mpg~hp+wt,data=mtcars)\\nsummary(fit)"
#read_pptx() \%>\% add_Rcode(code,title="Regression Analysis") \%>\% print(target="Rcode.pptx")
}
