% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_2ggplot2.R
\name{add_text}
\alias{add_text}
\title{Add text to document}
\usage{
add_text(mydoc, title = "", text, style = "Normal")
}
\arguments{
\item{mydoc}{A document object}

\item{title}{An character string as a plot title}

\item{text}{text string to be added}

\item{style}{text style}
}
\description{
Add text to document
}
