% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_2ggplot2.R
\name{add_title}
\alias{add_title}
\title{Add title to docx file}
\usage{
add_title(x, title = "", size = 20, color = NULL, before = TRUE,
  after = TRUE)
}
\arguments{
\item{x}{A document object}

\item{title}{Title}

\item{size}{font size}

\item{color}{font color}

\item{before}{Whether or not add blank paragraph before title}

\item{after}{Whether or not add blank paragraph after title}
}
\description{
Add title to docx file
}
