% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flextable.R
\name{add_title_slide}
\alias{add_title_slide}
\title{Add title slide}
\usage{
add_title_slide(mydoc, title = "", subtitle = "")
}
\arguments{
\item{mydoc}{A document object}

\item{title}{An character string as a title}

\item{subtitle}{An character string as a subtitle}
}
\description{
Add title slide
}
\examples{
#read_pptx() \%>\% add_title_slide(title="Web-based analysis with R" \%>\% print(target="title.pptx")
}
