% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptxList.R
\name{addFlexTableSlide}
\alias{addFlexTableSlide}
\title{Add a FlexTable slide into a document object}
\usage{
addFlexTableSlide(mydoc, mytable, title = "")
}
\arguments{
\item{mydoc}{A document object}

\item{mytable}{A FlexTable object}

\item{title}{An character string as a plot title}
}
\value{
a document object
}
\description{
Add a FlexTable slide into a document object
}
