% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptxList.R
\name{addggplot}
\alias{addggplot}
\title{Add ggplot into a document object}
\usage{
addggplot(mydoc, plot, title = "")
}
\arguments{
\item{mydoc}{A document object}

\item{plot}{An R code encoding a ggplot}

\item{title}{An character string as a plot title}
}
\value{
a document object
}
\description{
Add ggplot into a document object
}
