% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptxList.R
\name{addImageSlide}
\alias{addImageSlide}
\title{Add a plot into a document object}
\usage{
addImageSlide(mydoc, image, title = "")
}
\arguments{
\item{mydoc}{A document object}

\item{image}{Character vector indicating path to image file}

\item{title}{title of the slide}
}
\description{
Add a plot into a document object
}
