% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{addParagraphSlide}
\alias{addParagraphSlide}
\title{Add a paragraph slide into a document object}
\usage{
addParagraphSlide(mydoc, result = NULL, pre = NULL, post = NULL,
  title = "", fontsize = NULL, fontfamily = "Monaco", cat = FALSE,
  precat = TRUE, new = TRUE)
}
\arguments{
\item{mydoc}{A document object}

\item{result}{Character string encoding R codes as a main text}

\item{pre}{Character string encoding R codes as a preprocessing}

\item{post}{Character string encoding R codes as a postprocessing}

\item{title}{An character string as a plot title}

\item{fontsize}{A font size}

\item{fontfamily}{A character specifying the font family}

\item{cat}{A logical. If true, add line feed after main text}

\item{precat}{A logical. If true, add line feed before main text}

\item{new}{A logical. If TRUE, start with a new slide}
}
\value{
a document object
}
\description{
Add a paragraph slide into a document object
}
