% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{addplot}
\alias{addplot}
\title{Add plot into a document object}
\usage{
addplot(mydoc, plotfunction, title = "", vector = TRUE)
}
\arguments{
\item{mydoc}{A document object}

\item{plotfunction}{An R code encoding a plot}

\item{title}{An character string as a plot title}

\item{vector}{A logical. If TRUE, vector graphics are produced instead, PNG images if FALSE.}
}
\value{
a document object
}
\description{
Add plot into a document object
}
