% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{addRcodeSlide}
\alias{addRcodeSlide}
\title{Make a R code slide into a document object}
\usage{
addRcodeSlide(mydoc, code, title = "", showCode = FALSE,
  preprocessing = "")
}
\arguments{
\item{mydoc}{A document object}

\item{code}{A character string encoding R codes}

\item{title}{An character string as a plot title}

\item{showCode}{A logical. If true, show R code in a slide}

\item{preprocessing}{A character string of R code as a preprocessing}
}
\value{
a document object
}
\description{
Make a R code slide into a document object
}
