% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptxList.R
\name{addTitleSlide}
\alias{addTitleSlide}
\title{Add title slide into a document object}
\usage{
addTitleSlide(mydoc, title, adddate = TRUE, addsub = TRUE)
}
\arguments{
\item{mydoc}{A document object}

\item{title}{An character string as a plot title}

\item{adddate}{A logical. If TRUE, insert a date}

\item{addsub}{A logical. If TRUE, insert a subtitle}
}
\value{
a document object
}
\description{
Add title slide into a document object
}
