% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{canvas_cache}
\alias{canvas_cache}
\alias{canvas_destroy}
\alias{canvas_purge}
\title{Interact with the \code{OffscreenCanvas} cache}
\usage{
canvas_cache()

canvas_destroy(ids)

canvas_purge()
}
\arguments{
\item{ids}{Integer vector of canvas cache IDs.}
}
\description{
\code{canvas_cache()} returns an integer vector containing the current canvas
cache IDs.

\code{canvas_destroy()} destroys the cached \code{OffscreenCanvas} elements with the
given IDs.

\code{canvas_purge()} destroys all \code{OffscreenCanvas} elements in the cache.
}
