% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{canvas_install}
\alias{canvas_install}
\title{Use the webR canvas graphics device}
\usage{
canvas_install(...)
}
\arguments{
\item{...}{Arguments to be passed to the graphics device.}
}
\description{
Set R options so that the webR canvas graphics device is used as the default
graphics device for new plots.
}
