% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooserEx.R
\name{chooserInput}
\alias{chooserInput}
\title{Chooser Input}
\usage{
chooserInput(inputId, leftLabel, rightLabel, leftChoices, rightChoices,
  size = 5, multiple = FALSE, width = 100)
}
\arguments{
\item{inputId}{input Id}

\item{leftLabel}{Label for left column}

\item{rightLabel}{Label for right column}

\item{leftChoices}{choices for left column}

\item{rightChoices}{choices for right column}

\item{size}{number of column lines to be displayed}

\item{multiple}{logical enalble multiple selection}

\item{width}{width of left and right columns in pixel}
}
\description{
Chooser Input
}
