% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.stuart.test.R
\name{cox.stuart.test}
\alias{cox.stuart.test}
\title{Cox-Stuart test for trend analysis
The Cox-Stuart test is defined as a little powerful test (power equal to 0.78), but very robust for the trend analysis.
It is therefore applicable to a wide variety of situations, to get an idea of the evolution of values obtained.
The proposed method is based on the binomial distribution.
This function was written by Tommaso Martino<todoslogos@gmail.com> (See 'References')}
\usage{
cox.stuart.test(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A list with class "htest"
}
\description{
Cox-Stuart test for trend analysis
The Cox-Stuart test is defined as a little powerful test (power equal to 0.78), but very robust for the trend analysis.
It is therefore applicable to a wide variety of situations, to get an idea of the evolution of values obtained.
The proposed method is based on the binomial distribution.
This function was written by Tommaso Martino<todoslogos@gmail.com> (See 'References')
}
\examples{
customers = c(5, 9, 12, 18, 17, 16, 19, 20, 4, 3, 18, 16, 17, 15, 14)
cox.stuart.test(customers)
}
\references{
Original code: \url{http://statistic-on-air.blogspot.kr/2009/08/trend-analysis-with-cox-stuart-test-in.html}
}
