% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flextable.R
\name{data2docx}
\alias{data2docx}
\title{convert data to docx file}
\usage{
data2docx(...)
}
\arguments{
\item{...}{arguments to be passed to data2office()}
}
\description{
convert data to docx file
}
\examples{
#library(webr)
#data2docx(sampleData2)
}
