% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptxList.R
\name{data2HTML}
\alias{data2HTML}
\title{Make a HTML5 file with a data.frame}
\usage{
data2HTML(data, preprocessing = "", filename = "report.HTML",
  rawDataName = NULL, rawDataFile = "rawData.RDS", vanilla = FALSE,
  echo = FALSE)
}
\arguments{
\item{data}{A data.frame}

\item{preprocessing}{A character string of R code}

\item{filename}{A path of destination file}

\item{rawDataName}{The name of the rawData}

\item{rawDataFile}{The name of the rawData file which the data are to be read from.}

\item{vanilla}{logical. Whether or not make vanilla table}

\item{echo}{Logical. Whether or not show R code of plot and table}
}
\description{
Make a HTML5 file with a data.frame
}
\examples{
library(moonBook)
library(ztable)
library(webr)
data2HTML(sampleData2)
}
