% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flextable.R
\name{data2office}
\alias{data2office}
\title{convert data to pptx file}
\usage{
data2office(data, title = "Web-based Meta-Analysis", preprocessing = "",
  filename = "Report", format = "pptx", width = 7, height = 5,
  units = "in", res = 300, rawDataName = NULL,
  rawDataFile = "rawData.RDS")
}
\arguments{
\item{data}{A document object}

\item{title}{An character string as a title}

\item{preprocessing}{A string}

\item{filename}{File name}

\item{format}{desired format. choices are "pptx" or "docx"}

\item{width}{the width of the device.}

\item{height}{the height of the device.}

\item{units}{The units in which height and width are given. Can be px (pixels, the default), in (inches), cm or mm.}

\item{res}{The nominal resolution in ppi which will be recorded in the bitmap file, if a positive integer. Also used for units other than the default, and to convert points to pixels.}

\item{rawDataName}{raw Data Name}

\item{rawDataFile}{raw Data File}
}
\description{
convert data to pptx file
}
