% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{data2plotzip}
\alias{data2plotzip}
\title{Make zipped plot file with a data.frame}
\usage{
data2plotzip(data, filename = "Plot.zip", format = "PNG", width = 8,
  height = 6, units = "in", res = 300, start = 0, preprocessing = "")
}
\arguments{
\item{data}{A data.frame}

\item{filename}{A path of destination file}

\item{format}{Plot format. Choices are c("PNG","SVG","PDF")}

\item{width}{A plot width}

\item{height}{A plot height}

\item{units}{The units in which height and width are given. Can be px (pixels, the default), in (inches), cm or mm.}

\item{res}{The nominal resolution in ppi}

\item{start}{Plot start number}

\item{preprocessing}{A character string of R code}
}
\description{
Make zipped plot file with a data.frame
}
\examples{
type=c("mytable","table","plot","ggplot","Rcode")
title=c("mytable","iris[1:10,]","plot","ggplot","Regression Analysis")
code=c("mytable(sex~.,data=acs)","df2FlexTable(iris[1:10,])","plot(iris)","ggplot(iris,aes(x=Sepal.Length,y=Sepal.Width,color=Species))+geom_point()","summary(mtcars)\\nfit=lm(mpg~wt*hp,data=mtcars)\\nsummary(fit)")
data=data.frame(type=type,title=title,code=code,stringsAsFactors = FALSE)
data2plotzip(data)
}
