% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{data2pptx}
\alias{data2pptx}
\title{Make a powerpoint file with a data.frame}
\usage{
data2pptx(data, title = "Web-based Meta-Analysis", template = "myppt.pptx",
  preprocessing = "", filename = file, width = 7, height = 5,
  units = "in", res = 300, type = "cairo")
}
\arguments{
\item{data}{A data.frame}

\item{title}{The title of a powerpoint file}

\item{template}{Character value, it represents the filename of the pptx file used as a template.}

\item{preprocessing}{A character string of R code}

\item{filename}{A path of destination file}

\item{width}{A plot width}

\item{height}{A plot height}

\item{units}{The units in which height and width are given. Can be px (pixels, the default), in (inches), cm or mm.}

\item{res}{The nominal resolution in ppi}

\item{type}{A character}
}
\description{
Make a powerpoint file with a data.frame
}
\examples{
type=c("mytable","table","plot","ggplot","Rcode")
title=c("mytable","iris[1:10,]","plot","ggplot","Regression Analysis")
code=c("mytable(sex~.,data=acs)","df2FlexTable(iris[1:10,])","plot(iris)","ggplot(iris,aes(x=Sepal.Length,y=Sepal.Width,color=Species))+geom_point()","summary(mtcars)\\nfit=lm(mpg~wt*hp,data=mtcars)\\nsummary(fit)")
data=data.frame(type=type,title=title,code=code,stringsAsFactors = FALSE)
data2pptx(data)
}
