% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{data2word}
\alias{data2word}
\title{Make a word file with a data.frame}
\usage{
data2word(data, title, preprocessing = "", filename = "MetaReport.docx")
}
\arguments{
\item{data}{A data.frame}

\item{title}{The title of a word file}

\item{preprocessing}{A character string of R code}

\item{filename}{A path of destination file}
}
\description{
Make a word file with a data.frame
}
\examples{
type=c("mytable","table","plot","ggplot","Rcode")
title=c("mytable","iris[1:10,]","plot","ggplot","Regression Analysis")
code=c("mytable(sex~.,data=acs)","df2FlexTable(iris[1:10,])","plot(iris)","ggplot(iris,aes(x=Sepal.Length,y=Sepal.Width,color=Species))+geom_point()","summary(mtcars)\\nfit=lm(mpg~wt*hp,data=mtcars)\\nsummary(fit)")
data=data.frame(type=type,title=title,code=code,stringsAsFactors = FALSE)
data2word(data)
}
