% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_r}
\alias{eval_r}
\title{Evaluate the provided R code, call, or expression for webR}
\usage{
eval_r(
  expr,
  conditions = TRUE,
  streams = FALSE,
  autoprint = FALSE,
  handlers = TRUE,
  env = parent.frame()
)
}
\arguments{
\item{expr}{The R code, call or expression to evaluate.}

\item{conditions}{If \code{TRUE}, capture and return conditions raised during
execution.}

\item{streams}{If \code{TRUE}, capture and return the \code{stdout} and \code{stderr}
streams.}

\item{autoprint}{If \code{TRUE}, automatically print as if the expression was
written at an R console.}

\item{handlers}{If \code{TRUE}, execute using a \link{tryCatch} with handlers in place.}

\item{env}{The environment in which to evaluate.}
}
\description{
This function evaluates the provided R code, call, or expression with various
settings in place to configure behavior. The function is intended to be used
by the webR \code{evalR} API, rather than invoked directly by the end user.
}
