% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{evalRCode}
\alias{evalRCode}
\title{Evaluate R code for webR}
\usage{
evalRCode(
  code,
  conditions = TRUE,
  streams = FALSE,
  autoprint = FALSE,
  handlers = TRUE
)
}
\arguments{
\item{code}{The R code to evaluate.}

\item{conditions}{If `TRUE`, capture and return conditions raised during
execution.}

\item{streams}{If `TRUE`, capture and return the `stdout` and `stderr`
streams.}

\item{handlers}{If `TRUE`, execute the R code using a [tryCatch], with
handlers in place.}
}
\description{
This function evaluates the provided R code with various settings in place
to configure behavior. The function is intended to be used by the webR
`evalRCode` API, rather than invoked directly by the end user.
}
