% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptxList.R
\name{exportCSV}
\alias{exportCSV}
\title{Export pptList file to desired format}
\usage{
exportCSV(file, format = "HTML")
}
\arguments{
\item{file}{The name of the file which the data are to be read from.}

\item{format}{desired ouput format. Possible choices are one of the c("HTML","pdf","word","pptx","plotzip")}
}
\description{
Export pptList file to desired format
}
