% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqTable.R
\name{freqSummary}
\alias{freqSummary}
\title{Make table summarizing frequency}
\usage{
freqSummary(x, digits = 1, lang = "en")
}
\arguments{
\item{x}{A vector}

\item{digits}{integer indicating the number of decimal places}

\item{lang}{Language. choices are one of c("en","kor")}
}
\description{
Make table summarizing frequency
}
\examples{
require(moonBook)
freqSummary(acs$Dx)
freqSummary(acs$smoking,lang="kor")
}
