% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{GroupVar}
\alias{GroupVar}
\title{Extract categorical variables}
\usage{
GroupVar(df, max.ylev = 20)
}
\arguments{
\item{df}{a data.frame}

\item{max.ylev}{maximal length of unique values of catergorical variables}
}
\description{
Extract categorical variables
}
