% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_argument.R
\name{insert_argument}
\alias{insert_argument}
\title{replace argument of a function}
\usage{
insert_argument(code, argument, value)
}
\arguments{
\item{code}{string of function call}

\item{argument}{argument of function to be set}

\item{value}{value to be set}
}
\description{
replace argument of a function
}
