% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renew_dic.R
\name{langchoice1}
\alias{langchoice1}
\title{Select word}
\usage{
langchoice1(id, lang = "en")
}
\arguments{
\item{id}{data id}

\item{lang}{language. Possible choices are c("en","kor")}
}
\description{
Select word
}
