% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{makePot}
\alias{makePot}
\title{Make a formated text with R codes}
\usage{
makePot(result = NULL, pre = NULL, post = NULL, fontsize = NULL,
  fontfamily = "Monaco", cat = FALSE, precat = TRUE)
}
\arguments{
\item{result}{Character string encoding R codes as a main text}

\item{pre}{Character string encoding R codes as a preprocessing}

\item{post}{Character string encoding R codes as a postprocessing}

\item{fontsize}{A font size}

\item{fontfamily}{A character specifying the font family}

\item{cat}{A logical. If true, add line feed after main text}

\item{precat}{A logical. If true, add line feed before main text}
}
\description{
Make a formated text with R codes
}
