% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PieDonut.R
\name{makeSubColor}
\alias{makeSubColor}
\title{Make subcolors with main colors}
\usage{
makeSubColor(main, no = 3)
}
\arguments{
\item{main}{character. main colors}

\item{no}{number of subcolors}
}
\description{
Make subcolors with main colors
}
