% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mount.R
\name{mount}
\alias{mount}
\alias{unmount}
\title{Mount an Emscripten filesystem object}
\usage{
mount(mountpoint, source, type = "workerfs")

unmount(mountpoint)
}
\arguments{
\item{mountpoint}{a character string giving the path to a directory to mount
onto in the Emscripten virtual filesystem.}

\item{source}{a character string giving the location of the data source to be
mounted.}

\item{type}{a character string giving the type of Emscripten filesystem to be
mounted: "workerfs" or "nodefs".}
}
\description{
Uses the Emscripten filesystem API to mount a filesystem object onto a given
directory in the virtual filesystem. The mountpoint will be created if it
does not already exist.

When mounting an Emscripten "workerfs" type filesystem the \code{source} should
be the URL for a filesystem image with filename ending \code{.data}, as produced
by Emscripten's \code{file_packager} tool. The filesystem image and metadata will
be downloaded and mounted onto the directory \code{mountpoint}.

When mounting an Emscripten "nodefs" type filesystem, the \code{source} should be
the path to a physical directory on the host filesystem. The host directory
will be mapped into the virtual filesystem and mounted onto the directory
\code{mountpoint}. This filesystem type can only be used when webR is running
under Node.
}
