% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptxList.R
\name{myplot2}
\alias{myplot2}
\title{Make zipped plots with a data.frame}
\usage{
myplot2(data, format = "PNG", width = 7, height = 7, units = "in",
  res = 300, start = 0, preprocessing = "", rawDataName = NULL,
  rawDataFile = "rawData.RDS")
}
\arguments{
\item{data}{A data.frame}

\item{format}{Plot format. Choices are c("PNG","SVG","PDF")}

\item{width}{A plot width}

\item{height}{A plot height}

\item{units}{The units in which height and width are given. Can be px (pixels, the default), in (inches), cm or mm.}

\item{res}{The nominal resolution in ppi}

\item{start}{Plot start number}

\item{preprocessing}{A character string of R code}

\item{rawDataName}{The name of the rawData}

\item{rawDataFile}{The name of the rawData file which the data are to be read from.}
}
\description{
Make zipped plots with a data.frame
}
