% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytable2flextable.R
\name{mytable2flextable}
\alias{mytable2flextable}
\title{Convert mytable object to flextable}
\usage{
mytable2flextable(result, vanilla = TRUE, fontname = NULL, fontsize = 12)
}
\arguments{
\item{result}{An object of class "mytable"}

\item{vanilla}{A Logical.}

\item{fontname}{Font name}

\item{fontsize}{font size}
}
\description{
Convert mytable object to flextable
}
\examples{
require(moonBook)
result=mytable(smoking+Dx~.,data=acs)
mytable2flextable(result)
mytable2flextable(result,vanilla=FALSE)
result=mytable(Dx~.,data=acs)
mytable2flextable(result)
mytable2flextable(result,vanilla=FALSE)
}
