% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptxList.R
\name{mytable2FTable}
\alias{mytable2FTable}
\title{Make FlexTable with mytable : Internal function used in mytable2doc or mytable2ppt}
\usage{
mytable2FTable(res, vanilla = FALSE, bg = "#5B7778", parRight = TRUE,
  padding = 5, widths = NULL)
}
\arguments{
\item{res}{An object of class moonBook::mytable}

\item{vanilla}{A logical. If true, make a vanilla table}

\item{bg}{A background color}

\item{parRight}{A logical. If true, right text alignment}

\item{padding}{Paragraph left and right padding - 0 or positive integer value}

\item{widths}{A numeric vector specifying columns widths in inches.}
}
\value{
An object of ReporteRs::FlexTable
}
\description{
Make FlexTable with mytable : Internal function used in mytable2doc or mytable2ppt
}
\examples{
require(moonBook)
require(ReporteRs)
res=mytable(Dx~.,data=acs)
res=mytable(Dx+sex~.,data=acs)
mytable2FTable(res)

}
