% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pager.R
\name{pager_install}
\alias{pager_install}
\title{Generate an output message when file is displayed with the pager}
\usage{
pager_install()
}
\description{
When enabled, the R pager function is replaced so that a request to display
a file's contents instead generates a webR output message. The request is
forwarded to the main thread to be handled by the application loading webR.

This replaces the default R pager mechanism, which tries to start an external
pager binary using the \code{system()} function.
}
