% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{pastelf}
\alias{pastelf}
\title{Concatenate vectors after converting to character.}
\usage{
pastelf(...)
}
\arguments{
\item{...}{one or more R objects, to be converted to character vectors.}
}
\description{
Concatenate vectors after converting to character.
}
