% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{plotPDF2}
\alias{plotPDF2}
\title{Make pdf file with a plot code}
\usage{
plotPDF2(fun, file, width = 7, height = 5, units = "in", res = 300,
  ggplot = FALSE)
}
\arguments{
\item{fun}{A R code for plot}

\item{file}{A path of destination file}

\item{width}{A plot width}

\item{height}{A plot height}

\item{units}{The units in which height and width are given. Can be px (pixels, the default), in (inches), cm or mm.}

\item{res}{The nominal resolution in ppi}

\item{ggplot}{A logical. Set this argument true if th R code is for ggplot}
}
\description{
Make pdf file with a plot code
}
