% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptxList.R
\name{plotSVG2}
\alias{plotSVG2}
\title{Make SVG file with a plot code}
\usage{
plotSVG2(fun, file, width = 7, height = 7, ggplot = FALSE)
}
\arguments{
\item{fun}{A R code for plot}

\item{file}{A path of destination file}

\item{width}{A plot width}

\item{height}{A plot height}

\item{ggplot}{A logical. Set this argument true if th R code is for ggplot}
}
\description{
Make SVG file with a plot code
}
