% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptxList.R
\name{pptxList}
\alias{pptxList}
\title{Server function of pptxList shiny module}
\usage{
pptxList(input, output, session, data = reactive(""),
  preprocessing = reactive(""))
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{data}{A data object}

\item{preprocessing}{A character string of R code}
}
\description{
Server function of pptxList shiny module
}
