% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{pptxListInput}
\alias{pptxListInput}
\title{UI of pptxList shiny module}
\usage{
pptxListInput(id)
}
\arguments{
\item{id}{A string}
}
\description{
UI of pptxList shiny module
}
\examples{
library(shiny)
library(ggplot2)
library(ReporteRs)
library(editData)
library(moonBook)
library(readr)
ui=fluidPage(
    pptxListInput("pptxlist")
    )
server=function(input,output,session){
    mydf=callModule(pptxList,"pptxlist")
}
shinyApp(ui,server)
}
