% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flextable.R
\name{Rcode2df}
\alias{Rcode2df}
\title{Make a data.frame with character strings encoding R code}
\usage{
Rcode2df(result, preprocessing)
}
\arguments{
\item{result}{character strings encoding R code}

\item{preprocessing}{character strings encoding R code as a preprocessing}
}
\description{
Make a data.frame with character strings encoding R code
}
