% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flextable.R
\name{Rcode2flextable}
\alias{Rcode2flextable}
\title{Make a flextable object with character strings encoding R code}
\usage{
Rcode2flextable(result, preprocessing = "")
}
\arguments{
\item{result}{character strings encoding R code}

\item{preprocessing}{character strings encoding R code as a preprocessing}
}
\description{
Make a flextable object with character strings encoding R code
}
