% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_argument.R
\name{replace_argument}
\alias{replace_argument}
\title{replace argument of a function}
\usage{
replace_argument(substring, argument, value)
}
\arguments{
\item{substring}{string of function call}

\item{argument}{argument of function to be set}

\item{value}{value to be set}
}
\description{
replace argument of a function
}
