% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.test.R
\name{runs.test}
\alias{runs.test}
\title{Runs test for randomness}
\usage{
runs.test(y, plot.it = FALSE, alternative = c("two.sided",
  "positive.correlated", "negative.correlated"))
}
\arguments{
\item{y}{A vector}

\item{plot.it}{A logical. whether or not daw a plot}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}
}
\value{
A list with class "htest" containing the following components: statistic,p-value,method and data.name
}
\description{
Runs test for randomness
}
\examples{
y=c(1,2,2,1,1,2,1,2)
runs.test(y)
y=c("A","B","B","A","A","B","A","B")
runs.test(y,alternative="p")
}
