% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\docType{data}
\name{sampleData}
\alias{sampleData}
\title{Sample data for PPTxList
A dataset containing five objects for reproducible research}
\format{A data frame with 5 rows and three columns
\describe{
   \item{type}{type of data}
   \item{title}{title of data}
   \item{code}{R code of data}
}}
\usage{
sampleData
}
\description{
Sample data for PPTxList
A dataset containing five objects for reproducible research
}
\keyword{datasets}
