% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_argument.R
\name{set_argument}
\alias{set_argument}
\title{set argument of a function}
\usage{
set_argument(code, argument, value = TRUE)
}
\arguments{
\item{code}{string of function call}

\item{argument}{argument of function to be set}

\item{value}{value to be set}
}
\description{
set argument of a function
}
\examples{
code="df2flextable( ) "
code="df2flextable(vanilla=TRUE,head(iris[1:10,]))"
code="df2flextable(mtcars)"
code="df2flextable(head(iris[1:10,]),vanilla=TRUE)"
set_argument(code,"vanilla",FALSE)
}
