% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shims.R
\name{shim_install}
\alias{shim_install}
\title{Shim functions from base R}
\usage{
shim_install()
}
\description{
Replace base R functions with implementations
that work in the webR environment.
}
\details{
The shimmed functions are:
\itemize{
\item install.packages() - replaced by webr::install().
\item library() - replaced by webr::library_shim().
\item require() - replaced by webr::require_shim().
\item requireNamespace() - replaced by webr::requireNamespace_shim().
}
}
