% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mount.R
\name{syncfs}
\alias{syncfs}
\title{Synchronise the Emscripten virtual filesystem}
\usage{
syncfs(populate)
}
\arguments{
\item{populate}{A boolean. When \code{true}, initialises the filesystem with data
from persistent storage. When \code{false}, writes current filesystem data to
the persistent storage.}
}
\description{
Uses the Emscripten filesystem API to synchronise all mounted virtual
filesystems with their backing storage, where it exists. The \code{populate}
argument controls the direction of the synchronisation between Emscripten's
internal data and the file system's persistent store.
}
