% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flextable.R
\name{tensiSplit}
\alias{tensiSplit}
\title{Split strings with desired length with exdent}
\usage{
tensiSplit(string, size = 100, exdent = 3)
}
\arguments{
\item{string}{String}

\item{size}{desired length}

\item{exdent}{exdent}
}
\value{
splitted character vector
}
\description{
Split strings with desired length with exdent
}
