% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{test_package}
\alias{test_package}
\title{Test an installed R package by running the package examples and tests}
\usage{
test_package(pkg)
}
\arguments{
\item{pkg}{Name of the package to test.}
}
\value{
0 if the test was successful, otherwise 1.
}
\description{
This function runs R package examples and tests. The implementation is based
on \code{\link[tools:testInstalledPackage]{tools::testInstalledPackage()}}, with modifications for webR where the
\code{\link[=system]{system()}} function cannot be used.
}
