% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{viewer_install}
\alias{viewer_install}
\title{Generate an output message when a URL is browsed to}
\usage{
viewer_install()
}
\description{
When enabled, the R \code{viewer} option is set so that a request to display
a URL generates a webR output message. The request is forwarded to the main
thread to be handled by the application loading webR.

This does the equivalent of the base R function \code{utils::browseURL()}.
}
