% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webr-package.R
\docType{package}
\name{webr-package}
\alias{webr}
\alias{webr-package}
\title{webr: WebR Support Package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An R package to act as an interface to webR and provide supporting functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-wasm/webr}
  \item Report bugs at \url{https://github.com/r-wasm/webr/issues}
}

}
\author{
\strong{Maintainer}: George Stagg \email{george.stagg@posit.co}

Authors:
\itemize{
  \item Lionel Henry \email{lionel@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
