% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{writeCSVComment}
\alias{writeCSVComment}
\title{Write a cav file with comment}
\usage{
writeCSVComment(data, file, metadata = "", comment = "#")
}
\arguments{
\item{data}{A data.frame}

\item{file}{A path for destination file}

\item{metadata}{A character string representing R codes as a preprocessing}

\item{comment}{A string used to identify comments}
}
\description{
Write a cav file with comment
}
