% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x2Table.R
\name{x2summary}
\alias{x2summary}
\title{Summarize chisquare result}
\usage{
x2summary(
  data = NULL,
  x = NULL,
  y = NULL,
  a,
  b,
  margin = 1,
  show.percent = TRUE,
  show.label = TRUE
)
}
\arguments{
\item{data}{A data.frame}

\item{x}{a column name}

\item{y}{a column name}

\item{a}{a vector}

\item{b}{a vector}

\item{margin}{numeric If 1 row percent, if 2 col percent}

\item{show.percent}{logical}

\item{show.label}{logical}
}
\description{
Summarize chisquare result
}
\examples{
require(moonBook)
x2summary(acs,sex,DM)
}
