% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x2Table.R
\name{x2Table}
\alias{x2Table}
\title{Make a chisquare result table}
\usage{
x2Table(
  data,
  x,
  y,
  margin = 1,
  show.percent = TRUE,
  show.label = TRUE,
  show.stat = TRUE,
  vanilla = FALSE,
  fontsize = 12,
  ...
)
}
\arguments{
\item{data}{A data.frame}

\item{x}{a column name}

\item{y}{a column name}

\item{margin}{numeric If 1 row percent, if 2 col percent}

\item{show.percent}{logical}

\item{show.label}{logical}

\item{show.stat}{logical}

\item{vanilla}{logical whether or not make vanilla table}

\item{fontsize}{A numeric}

\item{...}{Further arguments to be passed to df2flextable()}
}
\description{
Make a chisquare result table
}
\examples{
require(moonBook)
x2Table(acs,sex,DM)
}
