% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_2ggplot2.R
\name{add_2ggplots}
\alias{add_2ggplots}
\title{Add two ggplots into a document object}
\usage{
add_2ggplots(mydoc, title = "Two plots", plot1, plot2, width = 3,
  height = 2.5, echo = FALSE)
}
\arguments{
\item{mydoc}{A document object}

\item{title}{An character string as a plot title}

\item{plot1}{An R code encoding the first ggplot}

\item{plot2}{An R code encoding the second ggplot}

\item{width}{plot width in inches}

\item{height}{plot height in inches}

\item{echo}{logical Whether or not show R code}
}
\value{
a document object
}
\description{
Add two ggplots into a document object
}
\examples{
require(ggplot2)
require(magrittr)
require(officer)
require(rvg)
plot1 <- "ggplot(data = iris, aes(Sepal.Length, Petal.Length)) + geom_point()"
plot2 <- "ggplot(data = iris, aes(Sepal.Length, Petal.Length, color = Species)) + geom_point()"
read_pptx() \%>\% add_2ggplots(title="Two plots",plot1=plot1,plot2=plot2,echo=TRUE)
}
