% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flextable.R
\name{add_flextable}
\alias{add_flextable}
\title{Add a flextable or mytable object into a document object}
\usage{
add_flextable(mydoc, ftable, title = "")
}
\arguments{
\item{mydoc}{A document object}

\item{ftable}{A flextable or mytable object}

\item{title}{An character string as a plot title}
}
\value{
a document object
}
\description{
Add a flextable or mytable object into a document object
}
\examples{
require(webr)
require(moonBook)
require(officer)
require(magrittr)
ftable=mytable(Dx~.,data=acs)
title="mytable Example"
ft=df2flextable(head(iris))
title2="df2flextable Example"
doc=read_docx()
doc \%>\% add_flextable(ftable,title) \%>\%
        add_flextable(ft,title2) \%>\%
        print(target="mytable.docx")
read_pptx() \%>\%
       add_flextable(ftable,title) \%>\%
       add_flextable(ft,title2) \%>\%
       print(target="mytable.pptx")
}
