% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flextable.R
\name{add_ggplot}
\alias{add_ggplot}
\title{Add ggplot into a document object}
\usage{
add_ggplot(mydoc, title = "", code = "", echo = FALSE)
}
\arguments{
\item{mydoc}{A document object}

\item{title}{An character string as a plot title}

\item{code}{R code for table}

\item{echo}{logical Whether or not show R code}
}
\value{
a document object
}
\description{
Add ggplot into a document object
}
\examples{
require(webr)
require(ggplot2)
require(officer)
require(magrittr)
code <- "ggplot(mtcars, aes(x = mpg , y = wt)) + geom_point()"
read_pptx() \%>\% add_ggplot(code=code,echo=TRUE)
}
