% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{canvas}
\alias{canvas}
\title{Graphics device for webR HTML canvas element}
\usage{
canvas(width = 504, height = 504, pointsize = 12, bg = "transparent", ...)
}
\arguments{
\item{width}{The width of the device.}

\item{height}{The height of the device.}

\item{pointsize}{The default point size of plotted text.}

\item{bg}{The initial background colour.}

\item{...}{Additional graphics device arguments (ignored).}
}
\description{
A graphics device that generates HTML canvas API calls and transmits them to
the main webR thread for plotting. The resulting messages are of type
\code{canvasExec}.
}
\details{
A 2x scaling is used to improve the bitmap output visual quality. As such,
the width and height of the HTML canvas element should be twice the width and
height of the graphics device.

Based on the \href{https://www.rforge.net/canvas/}{R canvas package} by
Jeffrey Horner, released under the GPL v2 Licence.
}
