% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{canvas_install}
\alias{canvas_install}
\title{Use the webR canvas device as the default graphics device}
\usage{
canvas_install(...)
}
\arguments{
\item{...}{Arguments to be passed to the canvas graphics device.}
}
\description{
Use the webR canvas device as the default graphics device
}
