% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{data2HTML}
\alias{data2HTML}
\title{Make a HTML5 file with a data.frame}
\usage{
data2HTML(data, preprocessing = "", filename = "report.HTML")
}
\arguments{
\item{data}{A data.frame}

\item{preprocessing}{A character string of R code}

\item{filename}{A path of destination file}
}
\description{
Make a HTML5 file with a data.frame
}
\examples{
library(moonBook)
library(ztable)
library(webr)
data2HTML(sampleData)
}
