% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{data2pdf}
\alias{data2pdf}
\title{Make a pdf file with a data.frame}
\usage{
data2pdf(data, preprocessing = "", filename = "report.pdf")
}
\arguments{
\item{data}{A data.frame}

\item{preprocessing}{A character string of R code}

\item{filename}{A path of destination file}
}
\description{
Make a pdf file with a data.frame
}
\examples{
library(moonBook)
library(ztable)
data2pdf(sampleData2)
}
