% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{data2pptx}
\alias{data2pptx}
\title{Make a powerpoint file with a data.frame}
\usage{
data2pptx(data, title = "Web-based Meta-Analysis", template = "myppt.pptx",
  preprocessing = "", filename = "Report.pptx", width = 7, height = 5,
  units = "in", res = 300)
}
\arguments{
\item{data}{A data.frame}

\item{title}{The title of a powerpoint file}

\item{template}{Character value, it represents the filename of the pptx file used as a template.}

\item{preprocessing}{A character string of R code}

\item{filename}{A path of destination file}

\item{width}{A plot width}

\item{height}{A plot height}

\item{units}{The units in which height and width are given. Can be px (pixels, the default), in (inches), cm or mm.}

\item{res}{The nominal resolution in ppi}
}
\description{
Make a powerpoint file with a data.frame
}
\examples{
library(moonBook)
library(ztable)
library(webr)
data2pptx(sampleData2)
}
