% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{data2word}
\alias{data2word}
\title{Make a word file with a data.frame}
\usage{
data2word(data, title, preprocessing = "", filename = "Report.docx",
  width = 7, height = 5, units = "in", res = 300)
}
\arguments{
\item{data}{A data.frame}

\item{title}{The title of a word file}

\item{preprocessing}{A character string of R code}

\item{filename}{A path of destination file}

\item{width}{A plot width}

\item{height}{A plot height}

\item{units}{The units in which height and width are given. Can be px (pixels, the default), in (inches), cm or mm.}

\item{res}{The nominal resolution in ppi}
}
\description{
Make a word file with a data.frame
}
\examples{
library(moonBook)
library(ztable)
library(webr)
data2word(sampleData)
}
