% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2FlexTable.R
\name{df2FlexTable}
\alias{df2FlexTable}
\title{Convert data.frame to FlexTable}
\usage{
df2FlexTable(df, vanilla = FALSE, bg = "#5B7778", add.rownames = TRUE,
  header.columns = TRUE, mode = "html", bordercolor = "#EDBD3E",
  oddcolor = "#FFFFFF", evencolor = "#DDDDDD", parRight = TRUE,
  parLeft = TRUE, padding.left = 5, padding.right = 5, padding.top = 2,
  padding.bottom = 2, widths = NULL, digits = NULL, NA2space = FALSE,
  pcol = NULL)
}
\arguments{
\item{df}{A data.frame}

\item{vanilla}{A logical. If true, make a vanilla table}

\item{bg}{A background color}

\item{add.rownames}{A logical. Whether or not add rownames}

\item{header.columns}{A logical. Whether or not add header columns}

\item{mode}{A character string. Choices are "html" or "latex"}

\item{bordercolor}{A border color}

\item{oddcolor}{A color of odd rows}

\item{evencolor}{A color of even rows}

\item{parRight}{A logical. If true, right text alignment}

\item{parLeft}{A logical. If true, left text alignment}

\item{padding.left}{Paragraph left padding - 0 or positive integer value}

\item{padding.right}{Paragraph right padding - 0 or positive integer value}

\item{padding.top}{Paragraph top padding - 0 or positive integer value}

\item{padding.bottom}{Paragraph bottom padding - 0 or positive integer value}

\item{widths}{A numeric vector specifying columns widths in inches.}

\item{digits}{An integer indicating the number of decimal places}

\item{NA2space}{A logical. If true, convert NA value to space}

\item{pcol}{An integer indicating p value. If specified, convert value less than 0.01 to "< 0.001" in given column.}
}
\value{
A FlexTable object
}
\description{
Convert data.frame to FlexTable
}
\examples{
library(ReporteRs)
df2FlexTable(head(iris))
df2FlexTable(head(iris),mode="pptx")
}
