% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flextable.R
\name{df2RcodeTable}
\alias{df2RcodeTable}
\title{Make a FlexTable with a data.frame}
\usage{
df2RcodeTable(df, bordercolor = "gray", maxlen = 80)
}
\arguments{
\item{df}{A data.frame}

\item{bordercolor}{A border color name}

\item{maxlen}{positive integer giving target line width in characters}
}
\value{
A FlexTable object
}
\description{
Make a FlexTable with a data.frame
}
