% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_js}
\alias{eval_js}
\title{Evaluate JavaScript code}
\usage{
eval_js(code)
}
\arguments{
\item{code}{The JavaScript code to evaluate.}
}
\value{
Integer result of evaluating the code.
}
\description{
This function evaluates the given character string as JavaScript code. The
result is returned as an integer.
}
\details{
The JavaScript code is evaluated using \code{emscripten_run_script_int} from the
Emscripten C API. In the event of a JavaScript exception an R error condition
will be raised with the exception message.

This is an experimental function that may undergo a breaking change in the
future so as to support different return types.
}
