% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_js}
\alias{eval_js}
\title{Evaluate JavaScript code}
\usage{
eval_js(code)
}
\arguments{
\item{code}{The JavaScript code to evaluate.}
}
\value{
Result of evaluating the JavaScript code, returned as an R object.
}
\description{
This function evaluates the given character string as JavaScript code.
Returned JavaScript objects are converted to R objects using the \code{RObject}
generic constructor, and specific R object types can be returned by invoking
the R object constructor directly in the evaluated JavaScript.
}
\details{
The JavaScript code is evaluated using \code{emscripten_run_script_int} from the
Emscripten C API. In the event of a JavaScript exception an R error condition
will be raised with the exception message.

This is an experimental function that may undergo a breaking changes in the
future.
}
\examples{
\dontrun{
eval_js("123 + 456")
eval_js("Math.sin(1)")
eval_js("true")
eval_js("undefined")
eval_js("(new Date()).toUTCString()")
eval_js("new RList({ foo: 123, bar: 456, baz: ['a', 'b', 'c']})")
}
}
