% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqTable.R
\name{freqTable}
\alias{freqTable}
\title{Make flextable summarizing frequency}
\usage{
freqTable(
  x,
  digits = 1,
  lang = getOption("freqTable.lang", "en"),
  vanilla = FALSE,
  ...
)
}
\arguments{
\item{x}{A vector}

\item{digits}{integer indicating the number of decimal places}

\item{lang}{Language. choices are one of c("en","kor")}

\item{vanilla}{Logical. Whether make vanilla table or not}

\item{...}{Further arguments to paseed to the df2flextable function}
}
\value{
An object of clss flextable
}
\description{
Make flextable summarizing frequency
}
\examples{
require(moonBook)
freqTable(acs$Dx)
#freqTable(acs$smoking,lang="kor",vanilla=TRUE,fontsize=12)
}
