% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{global_prompt_install}
\alias{global_prompt_install}
\title{Install a global handler to automatically download missing packages}
\usage{
global_prompt_install()
}
\description{
When enabled, \code{packageNotFoundError} errors invoke a global handler to
download missing R packages from the default webR binary repo.

If it is possible to do so, execution will continue without
interrupting the current program.
}
