% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{global_prompt_install}
\alias{global_prompt_install}
\title{Prompt user to install missing packages from a webR binary repo}
\usage{
global_prompt_install()
}
\description{
When enabled, `packageNotFoundError` errors cause a menu prompt
to be shown to the user asking if they would like to attempt to
download the missing package.

If the user replies in the negative, the package error is thrown as
it would normally. If the user replies in the affirmative, the package
is downloaded using [webr::install()].

If it is possible to do so, execution will continue without
interrupting the current program.
}
