% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install one or more packages from a webR binary package repo}
\usage{
install(packages, repos = NULL, lib = NULL)
}
\arguments{
\item{packages}{character vector containing the names of packages to install}

\item{repos}{character vector containing the URIs of the webR repos to use}

\item{lib}{the library directory where the packages will be installed}
}
\description{
Install one or more packages from a webR binary package repo
}
