% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install one or more packages from a webR binary package repo}
\usage{
install(packages, repos = NULL, lib = NULL, quiet = FALSE)
}
\arguments{
\item{packages}{Character vector containing the names of packages to install.}

\item{repos}{Character vector containing the URIs of the webR repos to use.}

\item{lib}{The library directory where the packages will be installed.}

\item{quiet}{Logical. If \code{TRUE}, do not output downloading messages.}
}
\description{
Install one or more packages from a webR binary package repo
}
