% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptxList.R
\name{Rcode2FlexTable}
\alias{Rcode2FlexTable}
\title{Make a FlexTable object with character strings encoding R code}
\usage{
Rcode2FlexTable(result, preprocessing = "")
}
\arguments{
\item{result}{character strings encoding R code}

\item{preprocessing}{character strings encoding R code as a preprocessing}
}
\description{
Make a FlexTable object with character strings encoding R code
}
