% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{readComment}
\alias{readComment}
\title{Read comment from a file}
\usage{
readComment(filename, comment = "#")
}
\arguments{
\item{filename}{A path for destination file}

\item{comment}{A string used to identify comments}
}
\description{
Read comment from a file
}
